WIP investigation of how to use -c
Essentially, it does not work the way we expect it to. The buildpack uses start.sh to set
up the environment and to run multiple processes (PHP, web server etc). This mechanism
is baked in at a low level and cannot easily be torn out.

This is the furthest we got. It finishes staging, runs the migration command, but does
not launch the webserver / PHP correctly.